refactor: 수신자 마음의 기록 파일 위치 모듈 컨벤션 정합
기존 모듈은 `viewmodel/` 에 VM·UiState 평탄화, `component/` 에 컴포넌트
평탄화, `screen/<role>/` 에 화면 파일만 두는 구조다. 신규 receiver 파일들을
같은 컨벤션으로 이동:

  screen/receiver/ReceiverMindRecordViewModel.kt   → viewmodel/
  screen/receiver/ReceiverMindRecordUiState.kt     → viewmodel/
  screen/receiver/component/CategoryChips.kt       → component/ReceiverCategoryChips.kt
  screen/receiver/component/Receiver*.kt           → component/

CategoryChips → ReceiverCategoryChips 로 함수명도 prefix 부여(component/ 평탄
구조에서 발신자측 카테고리 칩 등과 구분).

기능 변경 없음.
