+++ feature/mindrecord/presentation/src/main/kotlin/com/afternote/feature/mindrecord/presentation/viewmodel/ReceiverMindRecordUiState.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.mindrecord.presentation.screen.receiver
+package com.afternote.feature.mindrecord.presentation.viewmodel
 
 import com.afternote.feature.mindrecord.domain.model.MindRecordSummary
